Makefile tweaks and resolve #82
gx and gx-go will no longer reinstall if already installed, the phony rules have been added to the .PHONY list, and some of the logic has been simplified.

This could break things which manually use `make gxinstall`, `make testdeps`, or `make all`.
It's unlikely that anything uses `make all` because it is equivalent to `make` and the other two are just intermediary steps.