fixed hcdev gossip bug, resolves #343
problem was due to when NewNode was getting called in the gen sequence, it needs to be called after the DNA has is calculated and committed to the chain.  GenChain was calling NewNode (via Prepare) too early.

This fix also discovered that an error condition was not being returned, that might break some apps that didn't define a necessary genesis() function.